feat: AWS signature v4 lib
fix: remove unused code in test script
fix: signature V4 params

fix: canonical request params
