+++ oss/etc/nginx/conf.d/gateway/server_variables.conf
@@ -1,9 +1,9 @@
-# Variable indicating to the s3gateway.js script that singing key
+# Variable indicating to the awssig4.js script that singing key
 # caching is turned off. This feature uses the keyval store, so it
 # is only enabled when using NGINX Plus.
 set $cache_signing_key_enabled 0;
 
-# Variable indicating to the s3gateway.js script that session token
+# Variable indicating to the awscredentials.js script that session token
 # caching is turned on. This feature uses the keyval store, so it
 # is only enabled when using NGINX Plus.
 set $cache_instance_credentials_enabled 0;
