+++ standalone_ubuntu_oss_install.sh
@@ -350,6 +350,7 @@ EOF
 
 download "common/etc/nginx/include/listing.xsl" "/etc/nginx/include/listing.xsl"
 download "common/etc/nginx/include/awscredentials.js" "/etc/nginx/include/awscredentials.js"
+download "common/etc/nginx/include/awssig4.js" "/etc/nginx/include/awssig4.js"
 download "common/etc/nginx/include/s3gateway.js" "/etc/nginx/include/s3gateway.js"
 download "common/etc/nginx/include/utils.js" "/etc/nginx/include/utils.js"
 download "common/etc/nginx/templates/default.conf.template" "/etc/nginx/templates/default.conf.template"
